Multidimensional iterator; cleanup
A new iterator across Fortran column-major indices was added, named
FIndex.  This may be similar to dwwork's implementation.  The main
motivation was to write it as a Python iterator.

There is also some cleanup of dwwork's generalisation of the parse_index
function.

This is not a working commit!  Just an intermediate stage towards proper
multidimensional support.