Column-major data contiguity preserved in Python
Column-major multidimensional arrays in namelists are now converted to
(quasi-)row-major list of lists in Python.  This preserves data
contiguity between versions.

A `row_major` flag is provided to preserve the column-major syntax in
Python (at the cost of losing data contiguity).

The `append_value` function was moved inside the `Parser` class in order
to share the `row_major` configuration flag.