FIndex iteration fix to termination + test case
The column-major index iterator (FIndex) was not properly terminating,
and was instead looping over the final index.

We now allow the final value to increase unbounded, rather than
resetting it when it exceeds self.end[-1].  This allows us to test
agained self.end[-1] in the subsequent termination and terminate it if
necessary.

This was not previously noticed because we terminate iteration when all
of the requested values have been added, making bounds checking unused.

Further testing is probably required for FIndex, but everything is
working for now.