Patch case sensitivity bug fix
Previously, namelist group patches tested their (lowercase) group names
against the original case-insensitive group names of the original
namelist.

The bug was due to `get` not using the overrided `__getitem__` functions
of `Namelist`.

We now convert the Namelist group names to lowercase before calling
`get`, which should keep the problem under control.

A better longterm solution is to modify Namelist so that `get` is a
case-insensitive function (TODO?).

Thanks to github user ghislainp for this report.