Parsing floats with no `E` in the exponent.
Several Fortran compilers (gfortran, Intel, PGI) support floating point
data using E-notation without the 'E' but with a [+-] to denote sign and
separate mantissa and exponent.

We now check for this case and insert an 'e' if it is present.

Simple test cases have also been added.

Thanks to Jacob Williams for reporting this issue (#23) on github.