Add file arguments
Some API questions are how to handle `force=` and the default
output name for patching;  I decided to ignore `force` for
files.  For the latter issue, I raise `ValueError` when `nml_fname`
is a file since there is no sane default.

(personally, if it was my library, I would rip out the default name
 functionality entirely, and make it so that when there is no output
 file or path, it will still build and return the patched Namelist
 *object* without writing a patched namelist *file*.  But that's
 just me)