Namlelist patch output fix; partial list patch
Namelist patching was previously returning dicts whose values were
strings from `f90repr`, rather than the actual values.  This has been
fixed.

Patching lists was also causing an error, because the lists were being
passed through `f90repr`, which has been fixed in this patch.

However, patching is still causing errors in the patch output, since
only the first value of a vector was being replaced with the new values.
All other vector output was appended to the patched vector.

This patch is a partial fix: the original values are now suppressed, but
only the first vector name is written, the second vector name is not
printed.  (This is a side-effect of variable names being written before
they are parsed.)

This error is not reflected in the current tests, and needs to be fixed
ASAP.

Thanks to Sebastian Müller for reporting this problem.