Imperfect patch list name fix
We now manually re-write the variable name of patched lists, after
supressing it at the correct time.

The downside is that the whitespace is now out of order.  Any whitespace
after the variable name and the '=' token is written before the variable
name, rather than after.

THis is an imperfect fix to an imperfect implementation, but it will get
us through the day... remember to revist this.