Patch list bufgix
This is a more robust, but still temporary, fix for list patching.
We now keep write_token disable until we have written N-1 tokens.

This will only work if the original and patch variable list are equal
length.  This is generally a good assumption, since Fortran is not very
flexible with respect to arrays, but it could backfire someday.