New list patching
Patching was changed so that we now replace the patched token with the
new value, rather than suppressing the output to the patch file.

`update_tokens` was also modifiled with an optional token override input
argument.

This is a messy patch, but I don't want to lose the functional state.
The next patch will be a cleaned up version.