Bugfix in FIndex start and step iterations
Prior implemation was using a `if not b[0]` test to check for the
presence of a starting index, which reported false if `b[0]` was equal
to 0 (i.e. a zero-indexed array).

This patch replaces it with an explicit `if b[0] is None` test.

There was a similar problem with the strides, although a zero-stride is
probably not even defined in Fortran.