Improved (but incomplete) starting index support
Although starting index for vector ranges are correctly parsed and
provided, they are not being properly tracked and the translation to a
Python array typically assumes 1-based indexing.

This patch provides default values for the starting index stored in the
index iterator, which are used to build the array.

A single zero-based array assignment (`y(0:N) = 1, 2, ...`) will not
work correctly.

This is an incomplete patch which will break if one uses zero-indexing
but does not include the zero-index in the actual assignment of values
(which will still assume 1-indexing).

Global tracking of starting indices (probably within the namelist) will
be included in a subsequent patch.