Vector assigment after new starting index
Vector assignments with starting indexes lower than the original
(assumed) starting index will now correctly resize the vector and shift
the values to their new (Python) indices.

This is still not reflected in the `write` output, which will generally
not provide a 1-d index, and will give incorrect 1-based indexes for
higher dimensions.