Explicit index range output
The `write` method now includes the namelist index when outputting the
results, if the starting index has been set during `parse` or explicitly
by the user.  When no starting index is set, no indexing is assumed and
the output is displayed implcitly.

There may be some explicit 1-based indexing in higher dimensions of
multidimensional arrays if no information is available.

Numbering still starts at 1 even if no values are provided.  This needs
to be modified in `FIndex`.