Explicit start index; Implicit global start index
Previously, the starting index was very strongly implied to be 1 in most
cases.  For example, if we read the following variable:

    &grp
        x(3) = 1
    /

then it would be saved into Python as

    nml['grp']['x'] = [None, None, 1]

which may cause problems if the original Fortran vector was not
1-indexed (e.g. 0-indexed).

We now would save this as a single-valued vector:

    nml['grp']['x'] = [1]

but also track the starting index in a new field dict, `start_index`.

These are primarily used by the `write` method, which now begin printing
their output from the explicit starting index, rather than 1.

For example, the case above used to write the output as `x(1:3) = , , 1`
but now writes more simply as `x(3) = 1.

----

This has also exposed some bugs in defining the parents of children in
derived types.  In one case, a new child was not updating its status
with its parent.  This was previously not a problem, since the namelist
was empty and the parent re-establised the content.  But it needed to be
fixed since any local modifications to `start_index` were missed.  This
has now been fixed.

A second bug is that the child of a list of derived types incorrectly
sets its parent as the list of types, rather than the member which
spawned the child.  The bug seems to be related to the parsing of v_idx
(denoting the parent as a list, rather than the list member) and dtype
member on the same pass.  This is still **not fixed** but seems to have
no serious consequences yet.  Hopefully we can look into it someday in
the future.

----

There have also been some changes to implicit indexing, when a
variable's value is set in one stage and its index is set in another
stage, which is outlined below.

If we set a variable once with an index and again without an index, then
it's ambiguous where to place the non-indexed variable in the array.

This was not a serious problem when the starting index was very strongly
implied to be 1, but that is no longer the case.

We now used a `global_start_index` parameter, initially defined as 1, in
order to manage these situations.