Vector index output with no start index
The changes to make vectors agnostic to start index led to unexpected
behaviour.  In particular, the following record

    v = 1, 2, 3

would be output as

    v(:) = 1, 2, 3

which, if re-read and saved, would become

    v(1:3) = 1, 2, 3

which would incorrectly replace an unspecified initial index with 1.

We now preserve the initial state, so that unindexed vectors (`v = ...`)
are preserved, and vector index symbols without bounds (`v(:) = ...`)
are output as vectors without bounds.

This introduced one minor bug related to multidimensional arrays.  Since
the outer axes are explicitly expanded per-line, this can only be done
with an explicit index.  If none exists, then we *incorrectly* use
1-based indexing.  This reproduces the only behaviour, but is not
consistent with the other 1-based index improvements.

This bug cannot be fixed until we come up with a more elegant and
general method for writing multidimensional arrays to file, which is a
larger project.