Patch bugfix for vector of different size
Previously, the `patch` function only worked if the namelist and its
patch used vectors of the same size.  If the patch vector was smaller,
then the trailing values would still be retained in the patched namelist
file.

This patch fixes this issue by making the following changes:

1. Check if we are beyond the range of the patched vector; if so, then
   enable a new flag (`patch_skip`) in `update_tokens`.

2. Store patched whitespace and comments into a saved variable, rather
   than directly write to the patched file during token sweeping

3. When the new flag is on, check if the next value is either an old
   value or a new variable name.  This is done by checking if the next
   token is '='.

We also had to wrap one of the `next` calls around a StopIteration
exception check, and force an early write to patch if this occurs.  This
is because we rely on StopIteration to terminate the token sweep.
(Probably dumb but whatever...)

This was a mess to code, and probably indicates that f90nml has a lot of
crap that needs to be refactored ASAP.