Case-insensitive start index keys
The starting index keys saved in `start_index` of `Namelist` were
case-sensitive, which was conflicting with the (lowercase)
case-sensitive keys of `Namelist`.  This led to errors for keys with
uppercase characters.

This patch forces all keys inside `start_index` to be lowercase,
matching up with the `Namelist` keys.

There might be some subtler issues here that need investigating:

1. Why was a default index not assigned to the vectors with missing
   start index?

2. Should we try to preseve the case-sensitive variable names, even if
   Fortran does not respect them?