For derived types of lists, do not pass whole list
Previously, when a derived type parent was part of an indexed list (e.g.
`a%b(3)%c`) then we would pass the entire list `a%b` as the parent,
which required some gymnastics to resolve and also broke in a few cases
(seemed to be when an additional variable was introduced, but not sure.)

We now only pass the element `a%b(3)` as the parent, which appears to
resolve some of the newer cases.

(NOTE: The old list-checking hacks are still present, and the new test
cases have not been added.)