Locally incrase column width for very long keys
If keys exceed the colwidth of the namelist, then values were getting
removed during output.

We now locally increase colwidth to be the total header (+ indent)
length plus one, to ensure that one value will be written on each line.

This may not actually be the ideal or preferred behaviour for most
people, but it does fix the bug.