update_token whitespace check speedup
About 25% of update_token() time was spent checking for whitespace in
tokens using the all() builtin.  This was largely unnecessary, since the
new tokenizer always bundles whitespace into a single token, so that we
should only need to check the first token (as done with comments).

This new check led to a measureable speedup (about 0.9s on a 4.1s second
test, or about 20%).