Explicit copy of name tokens
We now scan the string explicitly for name token boundaries, copy the
word, and then manually replace the iterator with a new one which has
been advanced past the word.

This replaces the per-character iteration and inspection, which seems to
be notably faster (possibly 5-10%?).

This method is somewhat based on the comment tokenization, and should be
extendable to other tokens (strings, numbers).  This should also feed
back into the flint tokenizer.