Tokenization of multiline strings
The new Python-native tokenization was failing to tokenize multiline
strings.  The original multiline tokenizer from flint was not suitable
for namelists and was removed, but a working replacement was never
added.

The new version is slightly messy and reverts the amazing 1% speedup,
but now seems to work correctly.  Still being tested, and test cases
incoming.

Thanks for Robert Gieseke for reporting.