Bundled tokenization of strings outside of groups
Strings outside of namelist groups are (with some minor exceptions)
unhandled, and are predominantly used as comments.

Previously, f90nml was tokenizing this content as if it were valid
namelist syntax, which caused problems in some cases.

We now combine the entire line into a single token, as if it were a
comment.

There are a couple of syntatically relevant cases (specifically the
stdin/stdout markers) but those can be handled on another day.