Strong end comma support for long lines
Previously, we only checked for end commas for lines which were shorter
than the colwidth.  If a line is longer than a colwidth, it always had a
trailing comma.  This was because we presumed that only long lines with
multiple entries would exceed the colwidth.

For very short colwidths of a single entry, the comma was not being
respected.

We now inspect the final line of any entry, either long or short, and
remove the comma if unrequested.