CLI: start_index fix, patch test, __str__ fix
__str__() would fail for namelist groups, or anything that was not a
namelist file itself, a consequence of using Namelist() for everything.

We now only print the namelist as Fortran output for the file itself,
and print the ugly repr() for all contents.

There is room for improvement here, such as pretty-printing individual
groups, but needs a bit more care than I want to deal with right now.

----

We also now transfer start_index when using the Namelist constructor,
making it more "reflective" (as in `Namelist(self) = self`).  Still a
ways to go, but it's function now.

----

Finally, added a patch test case.