Saver multidimensional append_value updates
append_value would rename an input argument (v_values) when iterating
through its dimensional ranks (as list-of-list hierarchies).  No bugs
were apparent, but it was confusing the actual state and value of
v_values during execution.

We now use a different tag to indicate this recursive drill into the
list-of-list structure which (hopefully) reduces this confusion.