Manpage install via distutils
Since setuptools will not copy data files into explicit directories
(instead packing them into the egg) and since our only data file is a
manpage, we now use the explicit distutils command to install it the
"old way" to an explicit path.

This seems safe enough that I have enabled it on default (with the
directory safety checks).  So this still probably suffers a bit from
Unix-like assumptions.  But it works under both distutils and
setuptools.