CLI patch bugfix, doc improvements
Fixed a bug in the CLI with patches.  We should have been checking
output_fmt, not args.format, when confirming that the patched output is
a namelist.

Also, documentation was improved slightly.  Flags were formatted a bit
better (though they still look like shit in Sphinx) and a custom CSS was
added to prevent ligatures in code blocks.  (Almost wonder if this
should be standard in Sphinx??)