Dense arrays on default, sparse_arrays property
Dense arrays, with unknown values of multidimensional arrays replaced
with `None`, is now the default behaviour.  The `dense_arrays` property
was renamed to `sparse_arrays`.

Sparse arrays are not truly sparse for several reasons.  Later entries
appear to use the earlier entries to dictate the numer of None values
that must be applied.  This might be reduced to some extent.  But for
now, it moves the slightly confusing representation into a non-default
setting.