start_index property + docstring update
The Namelist start_index attribute is now stored as a property
(primarily to simplify the documentation.)

Much of the redundant documentation in the notes page was also removed,
since this information is now part of the API page.

The Namelist `write` method is now also documented in the API page.