List of dicts saved as Namelist; new type tests
The Namelist constructor was not converting lists of dicts into list of
Namelists, which was causing various problems related to indexing and
expectations of the start_index property.

There also was not very good support for arrays of dicts without a start
index.  This was presumed to not be a problem, since arrays of derived
types require an explicit index, but it was an issue when initialising
with a Python dict.

This patch fixes both issues.  Tests not yet included.

Thanks to GitHub user @gitporst for reporting.

----

We have also introduced the `is_nullable_list` function for checking
against list of lists or list of dicts, specifically for multidimensionl
array and derived type array tests.

Type testing in Namelist is also a bit stronger, since some functions do
presume to be an array of Namelists (due to start_index property) rather
than a more generic list of dicts.