Multidimension gen bugfix; Namelist gen tests
Similar to the previous bug related to generated Namelists from dicts of
dicts was a similar bug related to lists of lists (for multidimensional
arrays).  v_start indexing is now provisional, in case it is unset (as
with dict of dicts).  Additional code used to print a general array
index `:` for a null start index was enabled; previously there were no
examples which used this bit of code.

Additional tests for dict-of-dict and list-of-list dicts are now
included, although output is not tested.

Thanks to Github user @sound-fx for reporting.