default_start_index; Namelist recast update
A default_start_index attribute was added to the Namelist object and its
constructor.  This forces the start index to be a given value (e.g. 1)
and vector indices will always be printed.  The value is applied during
__setitem__ which is used by __init__.

Recasting of dict-like objects to Nameliset was moved from the
constructor to the __setitem__ method, which has cleaned up the code a
bit and should catch more (all?) of the possible cases.