setup.py test override
Older versions of setuptools (as in *much* older) were not picking up
the tests for the `test` command automatically.  So we now override the
`test` command with our own, regardless of whether or not we are using
setuptools or distutils.

This is probably a bit pedantic, but Travis 2.6 was not running the
tests, presumably because its setuptools was too old.  Since we went to
the trouble of manually writing a test command, we might as well use it
everywhere.