Handle vectors with values exceeding index
We now handle namelists with vectors with values outside the index.

Strictly I think these are interpreted as beyond the vector, and a new
entry, in which case they ought to be handled elsewhere.  But we can
formalise all this in a later version if necessary.

We ignore repeated commas, since these are essentialy null declarations.
But we raise warnings for unassined values and drop them from the
namelist.

Someday it might be worth adding some sort of strict flag which rejects
and terminates the namelist parsing.

Also added a test case for namelists without a teriminal CR (\n).