Bugfix: default_start_index value; private calls
default_start_index was returning the value of start_index, oops. Fixed.

Also, there were explicit calls to _f90repr and _var_strings in Parser,
which needed to be properly renamed.

The more relevant question is qhether these functions should be
re-labeled as public.

Fight for another day.  This patch fixes the bugs!