start_index and default_start_index bugfixes
This patch fixes some bugs related to setting the start_index and
default_start_index properties of Namelist.

First, we now parse args (via s_args) directly to check for start_index
when initialising the value.  We must do this prior to the __init__
call.  We must check args, since we next pass this to __init__ with the
_start_index keys removed.

We then call __init__, which expects and respects these values.  This
must be called after the start_index parsing.

Finally, we check and replace for any complex values inside __init__.
This must currently be done after init, since we manipulate the values
of `self` directly.  This should probably be changed someday to also
manipulate `args`, but not right now.

Also, a bug was fixed related to default_start_index, which was not
checking its value.  (It was checking the type of `self`, oops.)

Testing has also been slightly expanded.