Check for lists when using default_start_index
Namelist's default_start_index would force indexes to be displayed in
all cases, even if there is only a single value.  In this case, an index
was always displayed regardless of whether or not it was a list.

This was caused by our conversion of values to lists of one element
before constructing the index output.

To be honestly, the whole logic is a bit flawed, nor is this the only
issue, since we should not even be considering indices if the data is
not a list.  But not ready to fix this right now.

This fix restricts the use of default_start_index solely to lists of
data, and is part of the initial list/value check.

I am sure this whole logic could (and should) be streamlined.  But as
said before, that is work for another day.