Derived type lists with additional non-list values
If a derived type is set as a list, but subsequent scalar updates do not
include an index (i.e. if the update is the first element) then this was
raising an assertion.  However this can still correspond to a valid
namelist.  See the following example:

   &test
        var(1)%a = 1
        var%b    = 2
   /

Both var and var(1) correspond to the first element of a list of derived types.

This patch removes the assertion and forces the use of the default start
index (usually 1) if this occurs.

This issue possibly arised in part because derived type lists are
handled independently of vector lists, and could possibly be integrated
better.

Thanks to GitHub user @LSchueler for reporting this issue.