Index spacing after commas disabled on default
As explained in 10.11.2 of the standard, the namelist designator cannot
have any embedded blanks, e.g. `x(1,2)` is a valid designator but
`x(1, 2)` is not.

Because of this, index spacing is now disabled by default.  However,
since many compilers do support spaces, the feature will be retained for
now.

Thanks to Themo Tsakis for clarifying this issue.