String parser interface
We have added a function to the Parser object for parsing strings
containing namelist content, `reads`.  The string is broken into a
per-line iterator (as a file) and returns the Namelist object.

We have also set the default nml_patch argument of `_readstream` to
None, since it probably could do to be optional.  We may be able to also
eliminate some of the initialization of this variable in `Parser.read`.