Namelist indentation; zero-length indents
`Namelist` constructors now search for the `_indent` property and update
the indentation if this has been provided by the user.  This allows
generalized key-value fields to set indentation for the Python object.

This is primarily for use in patching, and allows `f90nml.patch` to
return namelists of different indentation.

Thanks to Maik Riechert for this suggestion.

We have also fixed a bug which supports zero-length indentation.
Previously, we checked for whitespace strings, which did not include
empty strings.