Multidimension out-of-order padding
Padding was not being correctly applied for multidimensional arrays when
entries were out of order, and starting index was decreasing as newer
entries were added to the variable.

In most cases the padding was being correctly counted, but was
re-applied again and again to the first dimension.  This is why many
cases with padding in the first dimension worked, but not more general
cases, which were missing from the test suite.

This patch has a more properly recursive method for sweeping the
variable and applying the padding.  There are still other bugs, but this
solves the specific issue of applying padding to the correct dimension.