README: Updated install support
Install instructions now direct users to use pip for local installs
rather than setup.py, e.g.

    pip install .

or

    pip install -e . --user

for non-root users.  This was due to some users having issues with the
classic setup.py installs, e.g.

    python setup.py install --user

I am not entirely sure what the preferred solution here is, but let's
try this one for a while and see the feedback.