Exclude Namelist in __setitem__ list of dicts
Previously, if given a list of dicts or Namelists, we could explictly
convert every (non-None) element into a Namelist, even if it was already
a namelist.  This is both redundant and recursive, since it would
subsequently apply this to any potential Namelist records within the
Namelist.

We now exlude Namelists from this conversion, under the assumption that
they would have been handled during the original conversion to Namelist.

This has improved the performance of namelists using lists of derived
types.  Thanks for Jacob Williams for reporting.