Array patching bugfix
This patch fixes two errors related to patching arrays of values.  The
first bug was in the remaining number of values in the array of the
unpatched namelist.  The `count_values` function did not include the
current token (self.token) but was also incorrectly counting multiple
whitespace tokens (e.g. '    ') as if it was not whitespace.

Typical indentation rules, and the fact that self.token is almost always
a valid value, was creating a scenario where these errors would cancel,
and typically give the correct answer, but would fail on certain edge
cases.

The iterator has been modified to include self.token and the whitespace
tracker now uses ''.isspace() to correctly identify whitespace tokens.

----

The second bug was due to the tracking of the current index of the patch
array (p_idx) and the number of remaining values to be patched in the
original array (n_vals_remain).

The original check (p_idx - 1) <= n_vals_remain (already odd in its
form) was both incrementing p_idx and decrementing n_vals_remain as we
iterated through the patch values, causing only about half of the vector
values to be updated, rather than all of them.

The patch replaces this with the correct and more intuitive check of
n_vals_remain to check that we do not attempt to apply values when the
patch is smaller than the original array.

----

This was likely unnoticed because all of our patch tests use very short
arrays of 1 or 2 values, which apparently did not fail in this case.

A longer array and patch of 5 elements has been added to the test suite
to check for this in the future.