Generalized comment support in tokenizer
The tokenizer had a hard-coded user-defined token (#) which is often
used for preprocessing but should not be taken as a standard comment
token.  This was user-defined in the parser but not in the tokenizer.

This patch replaces the fixed set with a field variable, which Parser
assigns before running `parse()`.