Tokenization streamline
This patch reduces many of the explicit tokenization tests for a much
simpler method.  Specifically, we no longer attempt to distinguish
between numerics, logicals, or variable names, which interfered with the
ability to identify non-delimited strings.

This shifts much of the work of identifying token type from the
tokenizer back onto the parser, which can identify them more robustly
via duck typing.

Although it seems that non-delimited strings are unsupported in the
Standard, they do work in Intel Fortran, and possibly others, so we
retain support for them here.  At this time, there's no warning or
indication of any problem with using them.