Bugfix: Repeat null value followed by comma
Repeated null values followed by a comma were incorrectly using the
comma as the repeat value, filling in the array with comma strings.
This was because were only detecting repeat tokens `*` followed by end
tokens (`/`, etc.).

This was resolved by looking for either * / or , / as the criterion for
using None as a value.

Thanks to github user @frankeye for reporting.