Improved null repeat support
Several null cases were still unhandled correctly, particularly if the
final record of a variable was a repeat null with a comma, and followed
by another variable.

After writing out all of the possible end-null cases, it seems that we
can safely conclude that the value is null when we have the following
token pairs (assuming that an earlier token was *)

* / :
    end of group, repeat null

<any> =:
    If the next token is an equals, then we have started the next
    assignment and the prior value was a repeat null.

, <any>:
    If the prior_token is a comma, then we had * , at an earlier
    point and this must be a repeat null

The special case of , , (two commas) seems to accidentally skip over a
null value, so we add one to n_vals in this case.  Subsequent nulls are
handled correctly, so we don't need to keep iteratively doing this.
There is just some problem with the iteration that we are overlooking.
But I have no plans to look into it.  Enjoy the random fix.

Test cases have also been added.

Again, thanks to GitHub user @frankeye for reporting this issue.