Integrate indent into output string header (p1)
This small change includes the indentation in any calculations of header
for handling multiline records.

This is a minor first step to applying any indentation for each record,
rather than prepending an indent for every record.

This will allow for indentation to be optional for some lines,
particularly for split strings.