Integrate indent into output string header (p3)
This patch completes the transfer of indent management to the variable string
output generation in the Namelist output.

column width is now explicit, rather than based on some offset from the
variable name, which simplifies the interpretation.  Several slightly
complex structures were also simplified due to the uniformity of most
operations.

The pathological case of writing output when the header exceeds the
columns is now handled slightly differently.  Previously, we were
looking for an empty string to avoid redundant writes.  Since we now
pre-pad the next potential record with indented whitespace, we must now
check for a whitespace string.