Raise ValueError on unhandled StopIteration
StopIteration is typically handled to catch a syntactically correct end
of token iteration.  But it is often still raised when there are syntax
errors and the end of file is unexpected.

This patch simply replaces the StopIteration exception with ValueError,
in order to at least consisently tag this event as an error.

A commented line from a previous patch has also been removed.