Test: Out-of-order dtype list, no start index
This adds a test for the supported but untested case of a list of
derived type elements, where the elements are out of order that the
first element is implied by the missing index, but is also not the first
element, e.g. [a(2), a, a(3)], where a is implicitly equal to a(1).