add compression routines
A new property added to namelist: compressed. Defaults to False.

If True, the output of the namelist is compressed by writing
successive identical values of arrays as n*v, where n is the number
of successive values and v is the value.

In order to do so, two new methods were added:

_compress: converts a list of values into a list of list of values,
e.g. [1, 1, 1, 2, 3, 3] -> [[3, 1], [1, 2], [2, 3]]

_f90comprepr: takes the elements of this compressed list and converts
them to their string components. In the above example:
"3*1", "2", and "2*3"

Call to these methods has been added to the _var_strings method