Preserve newlines in reads() string parser
The string parser function, reads(), used `splitlines` to produce a list
of per-line output, but also stripped the endline characters.

Normally the parser was robust enough to handle this, but it causes
isses for any tokens outside of the namelist groups, and tended to drop
the final character (presumably assuming it was an endline).

This patch adds a keepends flag to splitlines, which retains the endline
and keeps everything working.