Patch: Append additional namelists; indent bugfix
`f90nml.patch` will append additional variables to an existing namelist
group, but would not append new groups to an existing namelist.

This patch now gathers any remaining groups after a patch and now
appends them to the target namelist.

We also have fixed a bug where indentation was being applied twice to
patched values.  The indent is now part of the variable's string and no
longer needs to be prepended.

Thanks to Cory Kramer for reporting this bug.