Bugfix: Do not split string if next line fits
Several bugs had been introduced by the long string splitting methods
when applied to lists of strings.  Aside from arbitrarily splitting
smaller strings which would have previously remained on the current
line, redunant commas were also being inserted into split values.

Although the comma bug was resolved in the previous patch, this new
patch now attempts to avoid string splits when possible.

String splits are now more conditional.  If a string can fit on the next
line, we move it there.  If it cannot fit, then we apply a string split.

We also detected an error where trailing whitespace on the end of a
string split was dropped.  This whitespace is now preserved.

Thanks to Jared Lewis for reporting this error.