Key-based cogroup tracking
This change tracks the cogroup keys in a list, and checks against them
on every __getitem__ call.

The cogroup list is now generated on __getitem__ rather than
pre-computing it.  This avoids the messiness of updating the cogroup
list on every __setitem__.

Maybe I will regret these decisions, but they are the simplest path to
something functional.