Cogroup class
A Cogroup class was introduced to create a direct link between the
generated list of cogroups and the original values from the namelist.

Without this cogroup object, it is not possible to manipulate the list
of groups, which are served as a new list when created via __getitem__.

Individual elements can now be replaced via the access.

Much work to do, but this is at least starting to work as expected.