del support; fixed super() errors
del is now supported for cogroups via __delitem__ implementations.

Also removed some problematic super() usage; super() should only be used
to reference the parent class within a class.  We were using it to
access the parent class of a field in the class.  This has been fixed by
explicitly referencing dict.

(In truth, we should not even need super() and it should be sufficient
to explicit reference the parent, but oh well.)