Prevent Cogroup lists for Namelist(nml)
Currently, if we pass a namelist through `Namelist()`, it will normall
recreate itself.  This is not true for namelists containing cogroups,
which will return Cogroup lists for the prinicipal cogroup key.

For example::

    >>> nml = f90nml.reads('&grp x = 1 / &grp x = 2 /')
    >>> print(nml)
    &grp
        x = 1
    /

    &grp
        x = 2
    /
    >>> nml2 = f90nml.Namelist(nml)
    >>> print(nml2)
    &grp
        x = 1
    /

    &grp
        x = 2
    /

    &grp
        x = 2
    /
    >>>

This happens because nml['grp'] is returning a list when passed
internally through `super(Namelist, self).__init__()`.

We resolve this by "downgrading" a namelist back to an OrderedDict, and
then re-pass it through.

It is a bit redundant, but is an edge case for the most part and is
better than trying to reconstruct the cogroup values.

(This is more a conseqence of our desire to preserve the existing API,
and the difficult duality of treating the namelist simultaneously as a
list and a dict.)

We also fixed a bug where the `_cogroups` helper field was being reset.
It's now properly reconstructed from the values.