count_values stops after two valid tokens
The `count_values` function would count the number of available values
in a statement, in order to determine if a value was present and
assignable to the current variable.

However, this function was currently attempting to count the total
number of values, even though we only needed to determine if any values
existed.

For records with large numbers of values, this would attempt to count
all of the values, which is (unfortunately) an expensive operation.  For
example, an array of 8000 values would increase parsing time by a factor
of 100.

This patch modifies `count_values` to `check_for_values`, and now only
attempts to confirm that there is at least one value present.  If the
potential number of variables is at least two, then we terminate the
count.

For example::

    &grp x = 1 y = 2 /

would have a count of 2 after the first assignment (=): 1 and y.

We still retain the `ntoks > 0` check to count isolated values, e.g.::

    &grp x = 1 /

where the count is 1.

(Note that identifiers like `y` must be counted as values, since they
could indicate non-deliminted strings.)

Thanks to Stephen LeDoux for detecting and reporting this issue, and
Philip Cross for the proposed solution.  Marshall Ward made some minor
adjustments to the patch.

Commit log authored by Marshall Ward.

Co-authored-by: Marshall Ward <marshall.ward@gmail.com>