Cogroup set to dict; all cogroups use internal key
This patch redefines the self._cogroups to be a dict which tracks the
internal keys, rather than just a set of the public cogroup keys.

This patch also renames the initial cogroup to use the internal key name
(usually _grp_{name}_0) rather than just occupying the original key.
This ought to simplify the distinction between this value and the
cogroup as a whole.

Finally, `create_cogroup()` was introduced alongside add_cogroup(), to
separate the creation of the collection of cogroups from the addition of
a new cogroup.

Some other minor stuff was changed too, but will be summed up in the
proper commit log.  (e.g.  `default_start_index` is handled differently)